move GenerateContext to casadi_function_generation, formatting
